prepare data split into two parts
